# LogArtApp/backend/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: LogArt API
  version: 1.0.0
  description: Documentación API REST para la aplicación LogArt
  contact:
    name: David Moreno Martín
    email: david.moreno.m.w@gmail.com
    url: https://github.com/codeurjc-students/2024-logart
servers:
  - url: https://localhost:443/api/v1
    description: Servidor de desarrollo
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Comment:
      type: object
      properties:
        _id:
          type: string
          example: 60d0fe4f5311236168a109cd
        content:
          type: string
          example: Este es un comentario de ejemplo.
        object:
          type: string
          example: 60d0fe4f5311236168a109cc
        user:
          type: string
          example: 60d0fe4f5311236168a109ca
        isEditedByAdmin:
          type: boolean
          example: false
        editedBy:
          type: string
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2022-01-01T12:34:56Z'
        updatedAt:
          type: string
          format: date-time
          example: '2022-01-02T12:34:56Z'
    CommentUpdated:
      type: object
      properties:
        _id:
          type: string
          example: 60d0fe4f5311236168a109cd
        content:
          type: string
          example: Contenido actualizado del comentario.
        object:
          type: string
          example: 60d0fe4f5311236168a109cc
        user:
          type: string
          example: 60d0fe4f5311236168a109ca
        isEditedByAdmin:
          type: boolean
          example: false
        editedBy:
          type: string
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2022-01-01T12:34:56Z'
        updatedAt:
          type: string
          format: date-time
          example: '2022-01-02T12:34:56Z'
    CreateCommentRequest:
      type: object
      required:
        - content
        - objectId
      properties:
        content:
          type: string
          example: Este es un comentario de ejemplo.
        objectId:
          type: string
          example: 60d0fe4f5311236168a109cc
    CreateCommentResponse:
      type: object
      properties:
        comment:
          $ref: '#/components/schemas/Comment'
        message:
          type: string
          example: Comment created successfully
    UpdateCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: Contenido actualizado del comentario.
    UpdateCommentResponse:
      type: object
      properties:
        comment:
          $ref: '#/components/schemas/CommentUpdated'
        message:
          type: string
          example: Comment updated successfully
    DeleteCommentResponse:
      type: object
      properties:
        message:
          type: string
          example: Comment deleted successfully
    GetCommentsByObjectIdResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
      example:
        comments:
          - _id: 60d0fe4f5311236168a109cd
            content: Este es un comentario de ejemplo.
            object: 60d0fe4f5311236168a109cc
            user: 60d0fe4f5311236168a109ca
            isEditedByAdmin: false
            editedBy: null
            createdAt: '2022-01-01T12:34:56Z'
            updatedAt: '2022-01-02T12:34:56Z'
          - _id: 60d0fe4f5311236168a109ce
            content: Este es otro comentario de ejemplo editado por un admin.
            object: 60d0fe4f5311236168a109cc
            user: 60d0fe4f5311236168a109ca
            isEditedByAdmin: true
            editedBy: 60d0fe4f5311236168a109ca
            createdAt: '2022-01-02T12:34:56Z'
            updatedAt: '2022-01-03T12:34:56Z'
security:
  - BearerAuth: []
paths:
...
  /comments:
    post:
      summary: Crear un nuevo comentario
      tags:
        - Comentarios
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comentario creado exitosamente
          headers:
            Location:
              description: URL del nuevo comentario
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCommentResponse'
        '400':
          description: Error, solicitud incorrecta o datos faltantes
        '403':
          description: Error, no autorizado para comentar en este objeto
        '404':
          description: Error, objeto o usuario no encontrado
        '500':
          description: Error interno del servidor
  /comments/{commentId}:
    put:
      summary: Actualizar un comentario existente
      tags:
        - Comentarios
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: string
          description: ID del comentario a actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: Comentario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCommentResponse'
        '400':
          description: Error, solicitud incorrecta o formato de ID inválido
        '403':
          description: Error, no autorizado para actualizar este comentario
        '404':
          description: Error, comentario o usuario no encontrado
        '500':
          description: Error interno del servidor
    delete:
      summary: Eliminar un comentario por su ID
      tags:
        - Comentarios
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: string
          description: ID del comentario a eliminar
      responses:
        '200':
          description: Comentario eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCommentResponse'
        '400':
          description: Error, formato de ID de comentario inválido
        '403':
          description: Error, no autorizado para eliminar este comentario
        '404':
          description: Error, comentario o usuario no encontrado
        '500':
          description: Error interno del servidor
  /comments/{objectId}:
    get:
      summary: Obtener comentarios por ID de objeto
      tags:
        - Comentarios
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: objectId
          required: true
          schema:
            type: string
          description: ID del objeto para el cual se desean obtener los comentarios
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Número de página para paginación
        - in: query
          name: limit
          schema:
            type: integer
            default: 3
          description: Número de comentarios por página
        - in: query
          name: commentId
          schema:
            type: string
          description: ID específico de un comentario para obtener detalles
      responses:
        '200':
          description: Lista de comentarios obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCommentsByObjectIdResponse'
        '400':
          description: Error, formato de ID de objeto o comentario inválido
        '403':
          description: Error, no autorizado para ver comentarios de este objeto
        '404':
          description: Error, objeto o comentario no encontrado
        '500':
          description: Error interno del servidor
...